[bugfix] Fix pointer event propagation regression
