+++ src/engine/Events.ts
@@ -171,7 +171,8 @@ export class GameEvent<T, U = T> {
   public other: U | null;
 
   /**
-   * determines, if event bubbles to the target's ancestors
+   * If set to false, prevents event from propagating to other actors. If true it will be propagated
+   * to all actors that apply.
    */
   public bubbles: boolean = true;
