+++ src/engine/Input/Index.ts
@@ -9,5 +9,7 @@
  */
 export * from './Gamepad';
 export * from './Pointer';
+export * from './Pointers';
+export * from './PointerEvents';
 export * from './Keyboard';
 export * from './EngineInput';
