+++ src/engine/Interfaces/Index.ts
@@ -6,4 +6,4 @@ export * from './Evented';
 export * from './Loadable';
 export * from './Loader';
 export * from './LifecycleEvents';
-export * from './PointerEvents';
+export * from './PointerEventHandlers';
